+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/service/telqos/RecordCommand.java
@@ -15,22 +15,32 @@ import java.util.List;
 @TelqosCommand
 public class RecordCommand extends CliCommand {
 
+    /**
+     * @since 2.1.5
+     */
+    private static final String COMMAND_OPTION_STATUS = "status";
+
     private static final String COMMAND_OPTION_ONLINE = "online";
     private static final String COMMAND_OPTION_OFFLINE = "offline";
 
+
     private static final String[] COMMAND_OPTION_ARRAY = new String[]{
+            COMMAND_OPTION_STATUS,
             COMMAND_OPTION_ONLINE,
             COMMAND_OPTION_OFFLINE
     };
 
     private static final String IDENTITY = "record";
     private static final String DESCRIPTION = "记录功能上线/下线";
+    private static final String CMD_LINE_SYNTAX_STATUS = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_STATUS);
     private static final String CMD_LINE_SYNTAX_ONLINE = IDENTITY + " " +
             CommandUtil.concatOptionPrefix(COMMAND_OPTION_ONLINE);
     private static final String CMD_LINE_SYNTAX_OFFLINE = IDENTITY + " " +
             CommandUtil.concatOptionPrefix(COMMAND_OPTION_OFFLINE);
 
     private static final String[] CMD_LINE_ARRAY = new String[]{
+            CMD_LINE_SYNTAX_STATUS,
             CMD_LINE_SYNTAX_ONLINE,
             CMD_LINE_SYNTAX_OFFLINE
     };
@@ -47,6 +57,7 @@ public class RecordCommand extends CliCommand {
     @Override
     protected List<Option> buildOptions() {
         List<Option> list = new ArrayList<>();
+        list.add(Option.builder(COMMAND_OPTION_STATUS).desc("查看记录功能当前状态").build());
         list.add(Option.builder(COMMAND_OPTION_ONLINE).desc("上线服务").build());
         list.add(Option.builder(COMMAND_OPTION_OFFLINE).desc("下线服务").build());
         return list;
@@ -62,6 +73,9 @@ public class RecordCommand extends CliCommand {
                 return;
             }
             switch (pair.getLeft()) {
+                case COMMAND_OPTION_STATUS:
+                    context.sendMessage("记录功能当前状态: " + (recordQosService.isRecordStarted() ? "上线" : "下线"));
+                    break;
                 case COMMAND_OPTION_ONLINE:
                     recordQosService.startRecord();
                     context.sendMessage("记录功能已上线!");